equal spacing of buttons, extract disco debug
